requirements.txt
alembic==1.7.5
autopep8==1.6.0
Babel==2.9.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
dnspython==2.2.0
email-validator==1.1.3
Flask==2.0.2
Flask-Admin==1.5.8
Flask-BabelEx==0.9.4
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-Scss==0.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
mysqlclient==2.1.0
passlib==1.7.4
pycodestyle==2.8.0
pycparser==2.21
pyScss==1.3.7
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
speaklater==1.3
SQLAlchemy==1.4.29
toml==0.10.2
urllib3==1.26.8
Werkzeug==2.0.2
WTForms==3.0.1
